Method update
Indepenedent method for the choices include the left, right, middle, and short, long, and pass, rush options.  Still needs to the weighted variables for time and distance so that the results return increased risk/reward for distance and time.  Not all options are weighted for response correctly. suggest that the weights are determined by ratio,eg  2:1 (10, -5) is lower risk 1:2 is higher risk (5 , -10).